requirements.txt
attrs==23.1.0
bandit==1.7.5
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.1
cryptography==41.0.3
defusedxml==0.7.1
dparse==0.6.3
flake8==6.1.0
Flask==2.3.3
Flask-Cors==4.0.0
gitdb==4.0.10
GitPython==3.1.36
graphviz==0.20.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
numpy==1.25.2
packaging==21.3
pbr==5.11.1
pluggy==1.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.3.1
pytest-cov==4.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safety==2.3.5
smmap==5.0.1
stevedore==5.1.0
urllib3==2.0.4
Werkzeug==2.3.7
